chore(app): rename unit test and files
